+++ FastPriorityQueue.cs
@@ -100,7 +100,7 @@ namespace Priority_Queue
 #if NET_VERSION_4_5
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
-        public void Enqueue(T node, int priority)
+        public void Enqueue(T node, float priority)
         {
 #if DEBUG
             if (node == null)
@@ -311,7 +311,7 @@ namespace Priority_Queue
 #if NET_VERSION_4_5
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
-        public void UpdatePriority(T node, int priority)
+        public void UpdatePriority(T node, float priority)
         {
 #if DEBUG
             if (node == null)
